proyecto videojuegos

+++ proyecto-primer-bim/src/Principal.kt
@@ -2,18 +2,15 @@ import javax.swing.*
 import java.io.File
 import java.io.IOException
 
-
 val categorias: Array<Any> = arrayOf("Videojuegos", "Venta")
 val opciones: Array<Any> = arrayOf("Insertar", "Consultar", "Modificar", "Eliminar")
 
     fun main(args: Array<String>) {
-
-
         do {
             val result = JOptionPane.showOptionDialog(
                 null,
-                "Bienvenido al catálogo de videojuegos\nSeleccione una opción",
-                "Catálogo de Videojuegos",
+                "Bienvenido a la tienda de videojuegos\nSeleccione una opción",
+                "Tienda de Videojuegos",
                 1,
                 1,
                 null,
@@ -35,30 +32,34 @@ val opciones: Array<Any> = arrayOf("Insertar", "Consultar", "Modificar", "Elimin
                     when (result2) {
                         0 -> {
                             ingresarVideojuego()
-
                         }
                         1 -> {
-
                             val codJuego = JOptionPane.showInputDialog("Ingrese el Código del videojuego a buscar")
-                            consultarVideojuego(codJuego)
+                            if(codJuego!=null) {
+                                consultarVideojuego(codJuego)
+                            }
                         }
                         2 -> {
                             val codJuego =
                                 JOptionPane.showInputDialog("Ingrese el Código del videojuego a modificar")
-                            val nomJuego = JOptionPane.showInputDialog("Ingrese el Nombre del videojuego")
-                            val precioJuego = JOptionPane.showInputDialog("Ingrese el Precio del videojuego")
-                            val genJuego = JOptionPane.showInputDialog("Ingrese el Género del videojuego")
-                            val clasifJuego = JOptionPane.showInputDialog("Ingrese la Clasificación del videojuego")
-                            val platJuego = JOptionPane.showInputDialog("Ingrese la Plataforma del videojuego")
-                            val distJuego = JOptionPane.showInputDialog("Ingrese la Distribuidora del videojuego")
 
-                            JOptionPane.showMessageDialog(null, "Datos de Videojuego modificados")
+                            if(codJuego!=null) {
+                                modificarVideojuego(codJuego)
+                            }
                         }
                         3 -> {
                             val codJuego =
                                 JOptionPane.showInputDialog("Ingrese el Código del videojuego a eliminar")
+                            if(codJuego!=null) {
+                                if(consultarVideojuegoExiste(codJuego)!=null){
+                                    eliminarVideojuego(codJuego)
+                                    JOptionPane.showMessageDialog(null, "Videojuego eliminado")
+                                }
+                                else{
+                                    JOptionPane.showMessageDialog(null, "Videojuego no encontrado")
+                                }
+                            }
 
-                            eliminarVideojuego(codJuego)
                         }
 
                     }
@@ -76,32 +77,36 @@ val opciones: Array<Any> = arrayOf("Insertar", "Consultar", "Modificar", "Elimin
                     )
                     when (result2) {
                         0 -> {
-
                             ingresarVenta()
                         }
                         1 -> {
                             val codVenta = JOptionPane.showInputDialog("Ingrese el Código de la venta a buscar")
-                            consultarVenta(codVenta)
+                            if(codVenta!=null) {
+                                consultarVenta(codVenta)
+                            }
                         }
                         2 -> {
-
                             val codVenta = JOptionPane.showInputDialog("Ingrese el Código de la venta a modificar")
-                            val cedCliente = JOptionPane.showInputDialog("Ingrese el número de cédula del cliente")
-                            val fecha = JOptionPane.showInputDialog("Ingrese la fecha de la venta")
-                            val descripcion = JOptionPane.showInputDialog("Ingrese la descripción de la venta")
-                            val total = JOptionPane.showInputDialog("Ingrese el total de la venta")
 
-                            JOptionPane.showMessageDialog(null, "Datos de venta modificados")
+                            if(codVenta!=null) {
+                                modificarVenta(codVenta)
+                            }
                         }
                         3 -> {
                             val codVenta =
                                 JOptionPane.showInputDialog("Ingrese el Código de la venta a eliminar")
-                            eliminarVenta(codVenta)
+                            if(codVenta!=null) {
+                                if(consultarVentaExiste(codVenta)!=null){
+                                    eliminarVenta(codVenta)
+                                    JOptionPane.showMessageDialog(null, "Venta eliminada")
+                                }
+                                else{
+                                JOptionPane.showMessageDialog(null, "Venta no encontrada")
+                                }
+                            }
                         }
-
                     }
                 }
-
             }
         }while(result in 0..3)
     }
@@ -114,7 +119,7 @@ val opciones: Array<Any> = arrayOf("Insertar", "Consultar", "Modificar", "Elimin
 fun ingresarVideojuego() {
     val listaVideojuego = arrayListOf<Videojuego>()
         try {
-            var juego = Videojuego("null","null",0.0,"null","null","null","null")
+            val juego = Videojuego("null","null",0.0,"null","null","null","null")
             juego.codigoVideojuego = JOptionPane.showInputDialog("Ingrese el Código del videojuego")
             juego.nombreVideojuego = JOptionPane.showInputDialog("Ingrese el Nombre del videojuego")
             juego.precio = (JOptionPane.showInputDialog("Ingrese el Precio del videojuego")).toDouble()
@@ -122,7 +127,6 @@ fun ingresarVideojuego() {
             juego.clasificacion = JOptionPane.showInputDialog("Ingrese la Clasificación del videojuego")
             juego.plataforma = JOptionPane.showInputDialog("Ingrese la Plataforma del videojuego")
             juego.distribuidora = JOptionPane.showInputDialog("Ingrese la Distribuidora del videojuego")
-
 
             leerContenidoVideojuego("baseVideojuegos.txt")
             println(juego.toString())
@@ -143,35 +147,32 @@ fun ingresarVideojuego() {
 
 
 
-fun consultarListaVideojuegos() {
-    val listaVideojuego = arrayListOf<Videojuego>()
-    var juego: String = ""
-    for (item: Videojuego in listaVideojuego) {
-        juego += item.toString()
-    }
-    JOptionPane.showMessageDialog(null, juego)
-}
+
 
 fun consultarVideojuego(codigo: String){
     var registro = ""
     var cont = 0
-    consultarVideojuego2(codigo)?.forEach { it ->
-
-        registro +="Código: ${it.codigoVideojuego}\n"+"Nombre: ${it.nombreVideojuego}\n"+"Precio: ${it.precio}\n" +
+    val listaVideojuego = consultarVideojuegoExiste(codigo)
+    if(listaVideojuego!=null){
+        listaVideojuego.forEach {
+            registro +="Código: ${it.codigoVideojuego}\n"+"Nombre: ${it.nombreVideojuego}\n"+"Precio: ${it.precio}\n" +
                 "Género: ${it.genero}\n"+"Clasificación: ${it.clasificacion}\n"+"Plataforma: ${it.plataforma}\n" +
                 "Distribuidora: ${it.distribuidora}\n\n"
-        cont++
-    }
+            cont++
+        }
     JOptionPane.showMessageDialog(null,registro,"Resultado consulta",1,null)
-
     println(registro)
+    }
+    else{
+        JOptionPane.showMessageDialog(null, "Videojuego no encontrado")
+    }
 
 
 }
 
 
 
-fun consultarVideojuego2(codigo: String): List<Videojuego>?{
+fun consultarVideojuegoExiste(codigo: String): List<Videojuego>?{
     leerContenidoVideojuego("baseVideojuegos.txt").forEach {
         when(codigo){
             it.codigoVideojuego -> {
@@ -181,14 +182,13 @@ fun consultarVideojuego2(codigo: String): List<Videojuego>?{
             }
         }
     }
-
     return null
 }
 
 
 
 fun eliminarVideojuego(codigo: String){
-    val juegoElim = consultarVideojuego2(codigo)
+    val juegoElim = consultarVideojuegoExiste(codigo)
     val listaVideojuego = ArrayList<Videojuego>()
     if (juegoElim != null) {
         leerContenidoVideojuego("baseVideojuegos.txt").forEach {
@@ -197,20 +197,46 @@ fun eliminarVideojuego(codigo: String){
             }
         }
 
-        var registro = ""
         var text = ""
-        listaVideojuego.forEach { it ->
-            registro +="Código: ${it.codigoVideojuego}\n"+"Nombre: ${it.nombreVideojuego}\n"+"Precio: ${it.precio}\n" +
-                    "Género: ${it.genero}\n"+"Clasificación: ${it.clasificacion}\n"+"Plataforma: ${it.plataforma}\n" +
-                    "Distribuidora: ${it.distribuidora}\n\n"
+        listaVideojuego.forEach {
             text+="${it.codigoVideojuego},${it.nombreVideojuego},${it.precio},${it.genero},${it.clasificacion},${it.plataforma},${it.distribuidora}\n"
         }
 
         sobreescribirArchivo(text, "baseVideojuegos.txt")
+        leerContenidoVideojuego("baseVideojuegos.txt")
+    }
+    else{
+        JOptionPane.showMessageDialog(null,"Videojuego no encontrado")
+    }
+}
 
-        JOptionPane.showMessageDialog(null, "Videojuego eliminado")
+
+fun modificarVideojuego(codigo: String){
+
+    val listaVideojuego = consultarVideojuegoExiste(codigo)
+    var text = ""
+
+    if (listaVideojuego != null) {
+        eliminarVideojuego(codigo)
+
+        listaVideojuego[0].nombreVideojuego = JOptionPane.showInputDialog("Ingrese el Nombre del videojuego")
+        listaVideojuego[0].precio = (JOptionPane.showInputDialog("Ingrese el Precio del videojuego")).toDouble()
+        listaVideojuego[0].genero = JOptionPane.showInputDialog("Ingrese el Género del videojuego")
+        listaVideojuego[0].clasificacion = JOptionPane.showInputDialog("Ingrese la Clasificación del videojuego")
+        listaVideojuego[0].plataforma = JOptionPane.showInputDialog("Ingrese la Plataforma del videojuego")
+        listaVideojuego[0].distribuidora = JOptionPane.showInputDialog("Ingrese la Distribuidora del videojuego")
+
+        listaVideojuego.forEach {
+            text+="${it.codigoVideojuego},${it.nombreVideojuego},${it.precio},${it.genero},${it.clasificacion},${it.plataforma},${it.distribuidora}\n"
+        }
+
+        escribirArchivo(text, "baseVideojuegos.txt")
         leerContenidoVideojuego("baseVideojuegos.txt")
 
+        JOptionPane.showMessageDialog(null, "Datos de Videojuego modificados")
+    }
+    else{
+        JOptionPane.showMessageDialog(null,"Videojuego no encontrado")
     }
 }
 
@@ -231,12 +257,12 @@ fun ingresarVenta() {
         println(venta.toString())
         listaVenta.add(venta)
         escribirArchivo(venta.registroVenta(),"baseVentas.txt")
-        leerContenidoVideojuego("baseVentas.txt")
+        leerContenidoVenta("baseVentas.txt")
 
         JOptionPane.showMessageDialog(null, "Venta registrada")
 
     } catch (e: Exception){
-        //JOptionPane.showMessageDialog(null, "Error al ingresar datos")
+        JOptionPane.showMessageDialog(null, "Error al ingresar datos")
     }
 }
 
@@ -244,37 +270,34 @@ fun ingresarVenta() {
 
 
 
-fun consultarListaVentas() {
-    val listaVenta = arrayListOf<Venta>()
-    var venta: String = ""
-    for (item: Venta in listaVenta) {
-        venta += item.toString()
-    }
-    JOptionPane.showMessageDialog(null, venta)
-}
 
 fun consultarVenta(codigo: String){
     var registro = ""
     var cont = 0
-    consultarVenta2(codigo)?.forEach { it ->
+    val listaVenta = consultarVentaExiste(codigo)
+    if(listaVenta!=null){
 
+    listaVenta.forEach {
         registro +="Código: ${it.codigoVenta}\n"+"Cédula cliente: ${it.cedulaCliente}\n"+"Fecha de la venta: ${it.fechaVenta}\n" +
-                "Descripcióm: ${it.descripcion}\n"+"Total: ${it.total}\n\n"
+                "Descripción: ${it.descripcion}\n"+"Total: ${it.total}\n\n"
         cont++
     }
     JOptionPane.showMessageDialog(null,registro,"Resultado consulta",1,null)
 
-    println(registro)
-
+    //println(registro)
+    }
+    else{
+        JOptionPane.showMessageDialog(null, "Venta no encontrada")
+    }
 
 }
 
 
 
 
-fun consultarVenta2(codigo: String): List<Venta>?{
+fun consultarVentaExiste(codigo: String): List<Venta>?{
     leerContenidoVenta("baseVentas.txt").forEach {
-        when(codigo){
+        when (codigo) {
             it.codigoVenta -> {
                 return leerContenidoVenta("baseVentas.txt").filter { venta ->
                     venta.codigoVenta == codigo
@@ -282,14 +305,14 @@ fun consultarVenta2(codigo: String): List<Venta>?{
             }
         }
     }
-
     return null
 }
 
 
 
+
 fun eliminarVenta(codigo: String){
-    val ventaElim = consultarVenta2(codigo)
+    val ventaElim = consultarVentaExiste(codigo)
     val listaVenta = ArrayList<Venta>()
     if (ventaElim != null) {
         leerContenidoVenta("baseVentas.txt").forEach {
@@ -298,61 +321,46 @@ fun eliminarVenta(codigo: String){
             }
         }
 
-        var registro = ""
         var text = ""
-        listaVenta.forEach { it ->
-            registro +="Código: ${it.codigoVenta}\n"+"Cédula cliente: ${it.cedulaCliente}\n"+"Fecha de la venta: ${it.fechaVenta}\n" +
-                    "Descripcióm: ${it.descripcion}\n"+"Total: ${it.total}\n\n"
-
+        listaVenta.forEach {
             text+="${it.codigoVenta},${it.cedulaCliente},${it.fechaVenta},${it.descripcion},${it.total}\n"
         }
 
         sobreescribirArchivo(text, "baseVentas.txt")
-
-        JOptionPane.showMessageDialog(null, "Venta eliminada")
         leerContenidoVenta("baseVentas.txt")
-
+    }
+    else{
+        JOptionPane.showMessageDialog(null,"Venta no encontrada")
     }
 }
 
 
+fun modificarVenta(codigo: String){
 
+    val listaVenta = consultarVentaExiste(codigo)
+    var text = ""
 
+    if (listaVenta != null) {
+        eliminarVenta(codigo)
 
+        listaVenta[0].cedulaCliente = JOptionPane.showInputDialog("Ingrese el número de cédula del cliente")
+        listaVenta[0].fechaVenta = JOptionPane.showInputDialog("Ingrese la fecha de la venta")
+        listaVenta[0].descripcion = JOptionPane.showInputDialog("Ingrese la descripción de la venta")
+        listaVenta[0].total = (JOptionPane.showInputDialog("Ingrese el total de la venta")).toDouble()
 
+        listaVenta.forEach {
+            text+="${it.codigoVenta},${it.cedulaCliente},${it.fechaVenta},${it.descripcion},${it.total}\n"
+        }
 
+        escribirArchivo(text, "baseVentas.txt")
+        leerContenidoVenta("baseVentas.txt")
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+        JOptionPane.showMessageDialog(null, "Datos de venta modificados")
+    }
+    else{
+        JOptionPane.showMessageDialog(null,"Venta no encontrada")
+    }
+}
